@@ -16,8 +16,9 @@
   },
   "dependencies": {
     "color": "^3.0.0",
+    "date-fns": "^1.29.0",
     "email-validator": "^2.0.3",
-    "fp-ts": "^1.4.1",
+    "fp-ts": "^1.6.2",
     "io-ts": "^1.1.4",
     "italia-ts-commons": "^2.0.0",
     "lodash": "^4.17.10",
@@ -34,6 +35,7 @@
     "react-native-mixpanel": "^0.0.16",
     "react-native-modal": "^6.0.0",
     "react-native-offline": "^3.7.4",
+    "react-native-popup-menu": "^0.13.1",
     "react-native-push-notification": "^3.0.2",
     "react-native-sha256": "^1.0.1",
     "react-native-vector-icons": "^4.6.0",
@@ -43,9 +45,8 @@
     "redux": "4.0.0",
     "redux-form": "^7.3.0",
     "redux-logger": "3.0.6",
-    "redux-persist": "5.9.1",
+    "redux-persist": "5.10.0",
     "redux-saga": "^0.16.0",
-    "redux-thunk": "2.2.0",
     "reselect": "^3.0.1",
     "source-map": "^0.6.1",
     "stacktrace-js": "^2.0.0",
@@ -67,7 +68,7 @@
     "@types/react-navigation": "^1.5.10",
     "@types/react-redux": "^5.0.18",
     "@types/react-test-renderer": "^16.0.1",
-    "@types/redux-form": "^7.2.4",
+    "@types/redux-form": "^7.2.6",
     "@types/redux-logger": "^3.0.6",
     "@types/redux-saga": "^0.10.5",
     "@types/stacktrace-js": "^0.0.32",
@@ -75,7 +76,7 @@
     "@types/validator": "^9.4.1",
     "babel-jest": "23.0.1",
     "babel-preset-react-native": "4.0.0",
-    "danger": "^3.6.0",
+    "danger": "^3.7.18",
     "danger-plugin-digitalcitizenship": "^0.2.0",
     "italia-utils": "^3.6.1",
     "jest": "23.0.1",
@@ -86,12 +87,12 @@
     "ts-jest": "^22.4.6",
     "tslint": "^5.10.0",
     "tslint-config-prettier": "^1.13.0",
-    "tslint-eslint-rules": "^5.1.0",
-    "tslint-immutable": "^4.5.4",
+    "tslint-eslint-rules": "^5.3.1",
+    "tslint-immutable": "^4.6.0",
     "tslint-plugin-prettier": "^1.3.0",
     "tslint-react": "^3.5.1",
     "tslint-sonarts": "^1.7.0",
-    "typescript": "^2.8.3",
+    "typescript": "^2.9.1",
     "typestrict": "^0.0.8"
   },
   "jest": {
