@@ -5,24 +5,22 @@
 import * as React from "react";
 import I18n from "../../i18n";
 
-import { Content, View } from "native-base";
-import { Image } from "react-native";
+import { Text, View } from "native-base";
 import {
   NavigationInjectedProps,
   NavigationScreenProp,
   NavigationState
 } from "react-navigation";
 
-import { WalletAPI } from "../../api/wallet/wallet-api";
+import { connect } from "react-redux";
 import { WalletStyles } from "../../components/styles/wallet";
-import { WalletLayout } from "../../components/wallet/layout/WalletLayout";
-import { TransactionsList } from "../../components/wallet/TransactionsList";
-import { CreditCard } from "../../types/CreditCard";
-import { WalletTransaction } from "../../types/wallet";
-
-import { topContentTouchable } from "../../components/wallet/layout/types";
-
-const cardsImage = require("../../../img/wallet/card-tab.png");
+import TransactionsList, {
+  TransactionsDisplayed
+} from "../../components/wallet/TransactionsList";
+import { CardEnum, WalletLayout } from "../../components/wallet/WalletLayout";
+import { GlobalState } from "../../store/reducers/types";
+import { selectedCreditCardSelector } from "../../store/reducers/wallet/cards";
+import { CreditCard, UNKNOWN_CARD } from "../../types/CreditCard";
 
 interface ParamType {
   readonly card: CreditCard;
@@ -36,47 +34,45 @@ interface OwnProps {
   readonly navigation: NavigationScreenProp<StateParams>;
 }
 
-type Props = OwnProps & NavigationInjectedProps;
+type ReduxMappedProps = Readonly<{
+  selectedCard: CreditCard;
+}>;
 
-export class TransactionsScreen extends React.Component<Props, never> {
-  private touchableContent(): React.ReactElement<any> {
-    // TODO: change this with an actual component @https://www.pivotaltracker.com/story/show/157422715
-    return (
-      <View style={WalletStyles.container}>
-        <Image
-          style={WalletStyles.pfTabCard}
-          source={cardsImage}
-          resizeMode="contain"
-        />
+type Props = ReduxMappedProps & OwnProps & NavigationInjectedProps;
+
+class TransactionsScreen extends React.Component<Props, never> {
+  public render(): React.ReactNode {
+    const headerContents = (
+      <View>
+        <View style={WalletStyles.walletBannerText}>
+          <Text style={WalletStyles.white}>
+            {I18n.t("wallet.creditDebitCards")}
+          </Text>
+        </View>
+        <View spacer={true} />
       </View>
     );
-  }
-
-  public render(): React.ReactNode {
-    const card: CreditCard = this.props.navigation.state.params.card;
-    const transactions: ReadonlyArray<
-      WalletTransaction
-    > = WalletAPI.getTransactions(card.id);
-
-    const topContent = topContentTouchable(this.touchableContent());
 
     return (
       <WalletLayout
-        headerTitle={I18n.t("wallet.paymentMethod")}
-        allowGoBack={true}
+        title={I18n.t("wallet.paymentMethod")}
         navigation={this.props.navigation}
-        title={I18n.t("wallet.creditDebitCards")}
-        topContent={topContent}
+        showPayButton={false}
+        headerContents={headerContents}
+        cardType={{ type: CardEnum.FULL, card: this.props.selectedCard }}
       >
-        <Content style={WalletStyles.whiteContent}>
-          <TransactionsList
-            title={I18n.t("wallet.transactions")}
-            totalAmount={I18n.t("wallet.total")}
-            transactions={transactions}
-            navigation={this.props.navigation}
-          />
-        </Content>
+        <TransactionsList
+          title={I18n.t("wallet.transactions")}
+          totalAmount={I18n.t("wallet.total")}
+          navigation={this.props.navigation}
+          display={TransactionsDisplayed.BY_CARD}
+        />
       </WalletLayout>
     );
   }
 }
+
+const mapStateToProps = (state: GlobalState): ReduxMappedProps => ({
+  selectedCard: selectedCreditCardSelector(state).getOrElse(UNKNOWN_CARD)
+});
+export default connect(mapStateToProps)(TransactionsScreen);
