@@ -11,7 +11,7 @@ import I18n from "../../i18n";
 import { WalletTransaction } from "../../types/wallet";
 
 import { NonEmptyString } from "italia-ts-commons/lib/strings";
-import { CreditCard, UNKNOWN_CARD } from "../../types/CreditCard";
+import { CreditCard } from "../../types/CreditCard";
 
 // temporarily making this a variable
 // (to mock the deleteCreditCard() api more easily)
@@ -57,9 +57,11 @@ const cards: ReadonlyArray<CreditCard> = [
 
 const transactions: ReadonlyArray<WalletTransaction> = [
   {
+    id: 1,
     cardId: 1,
     date: "17/04/2018",
     time: "07:34",
+    isoDatetime: "2018-04-17T07:34:00.000Z",
     paymentReason: "Certificato di residenza",
     recipient: "Comune di Gallarate",
     amount: -20.02,
@@ -68,9 +70,11 @@ const transactions: ReadonlyArray<WalletTransaction> = [
     isNew: true
   },
   {
+    id: 2,
     cardId: 2,
     date: "16/04/2018",
     time: "15:01",
+    isoDatetime: "2018-04-16T15:01:00.000Z",
     paymentReason: "Spesa Supermarket",
     recipient: "Segrate",
     amount: -74.1,
@@ -79,9 +83,11 @@ const transactions: ReadonlyArray<WalletTransaction> = [
     isNew: true
   },
   {
+    id: 3,
     cardId: 4,
     date: "15/04/2018",
     time: "08:56",
+    isoDatetime: "2018-04-15T08:56:00.000Z",
     paymentReason: "Prelievo contante",
     recipient: "Busto Arsizio",
     amount: -200.0,
@@ -90,9 +96,11 @@ const transactions: ReadonlyArray<WalletTransaction> = [
     isNew: true
   },
   {
+    id: 4,
     cardId: 2,
     date: "14/02/2018",
     time: "10:21",
+    isoDatetime: "2018-02-14T10:21:00.000Z",
     paymentReason: "Accredito per storno",
     recipient: "Banca Sella",
     amount: 100.1,
@@ -101,9 +109,11 @@ const transactions: ReadonlyArray<WalletTransaction> = [
     isNew: false
   },
   {
+    id: 5,
     cardId: 4,
     date: "22/01/2018",
     time: "14:54",
+    isoDatetime: "2018-01-22T14:54:00.000Z",
     paymentReason: "Esecuzione atti notarili",
     recipient: "Comune di Legnano",
     transactionCost: 0.5,
@@ -112,9 +122,11 @@ const transactions: ReadonlyArray<WalletTransaction> = [
     isNew: false
   },
   {
+    id: 6,
     cardId: 4,
     date: "01/01/2018",
     time: "23:34",
+    isoDatetime: "2018-01-01T23:34:00.000Z",
     paymentReason: "Pizzeria Da Gennarino",
     recipient: "Busto Arsizio",
     amount: -45.0,
@@ -123,9 +135,11 @@ const transactions: ReadonlyArray<WalletTransaction> = [
     isNew: false
   },
   {
+    id: 7,
     cardId: 1,
     date: "22/12/2017",
     time: "14:23",
+    isoDatetime: "2017-12-22T14:23:00.000Z",
     paymentReason: "Rimborso TARI 2012",
     recipient: "Comune di Gallarate",
     amount: 150.2,
@@ -134,9 +148,11 @@ const transactions: ReadonlyArray<WalletTransaction> = [
     isNew: false
   },
   {
+    id: 8,
     cardId: 1,
     date: "17/12/2017",
     time: "12:34",
+    isoDatetime: "2017-12-17T12:34:00.000Z",
     paymentReason: "Ristorante I Pini",
     recipient: "Busto Arsizio",
     transactionCost: 0,
@@ -145,9 +161,11 @@ const transactions: ReadonlyArray<WalletTransaction> = [
     isNew: false
   },
   {
+    id: 9,
     cardId: 4,
     date: "13/12/2017",
     time: "10:34",
+    isoDatetime: "2017-12-13T10:34:00.000Z",
     paymentReason: "Estetista Estella",
     recipient: "Milano - via Parini 12",
     transactionCost: 0.5,
@@ -165,31 +183,13 @@ const transactions: ReadonlyArray<WalletTransaction> = [
 // will actually fetch data from the proxy/pagopa)
 // @https://www.pivotaltracker.com/story/show/157770129
 export class WalletAPI {
-  public static readonly MAX_TRANSACTIONS = 5;
-
-  public static getCreditCards(): ReadonlyArray<CreditCard> {
+  public static async getCreditCards(): Promise<ReadonlyArray<CreditCard>> {
     return cards;
   }
 
-  public static getCreditCard(creditCardId: number): CreditCard {
-    const card = cards.find((c: CreditCard): boolean => c.id === creditCardId);
-    if (card === undefined) {
-      return UNKNOWN_CARD;
-    }
-    return card;
-  }
-
-  public static getTransactions(
-    cardId: number
-  ): ReadonlyArray<WalletTransaction> {
-    return transactions.filter(
-      (transaction): boolean => transaction.cardId === cardId
-    );
-  }
-
-  public static getLatestTransactions(
-    maxOps: number = WalletAPI.MAX_TRANSACTIONS
-  ): ReadonlyArray<WalletTransaction> {
-    return transactions.slice(0, maxOps);
+  public static async getTransactions(): Promise<
+    ReadonlyArray<WalletTransaction>
+  > {
+    return transactions;
   }
 }
