@@ -1,37 +1,37 @@
+import { AsyncStorage } from "react-native";
 import { NavigationState } from "react-navigation";
 import { createReactNavigationReduxMiddleware } from "react-navigation-redux-helpers";
 import { applyMiddleware, compose, createStore, Reducer } from "redux";
 import { createLogger } from "redux-logger";
 import {
-  persistCombineReducers,
   PersistConfig,
   Persistor,
+  persistReducer,
   persistStore
 } from "redux-persist";
-import storage from "redux-persist/lib/storage";
 import createSagaMiddleware from "redux-saga";
-import thunk from "redux-thunk";
 
 import { analytics } from "../middlewares";
-import rootReducer from "../reducers";
-import { GlobalState } from "../reducers/types";
 import rootSaga from "../sagas";
 import { Action, Store, StoreEnhancer } from "../store/actions/types";
+import rootReducer from "../store/reducers";
+import { GlobalState, PersistedGlobalState } from "../store/reducers/types";
 import { NAVIGATION_MIDDLEWARE_LISTENERS_KEY } from "../utils/constants";
 
 const isDebuggingInChrome = __DEV__ && !!window.navigator.userAgent;
 
 const persistConfig: PersistConfig = {
   key: "root",
-  storage,
+  storage: AsyncStorage,
   /**
-   * Sections of the store that must be persisted and rehydrated.
+   * Sections of the store that must be persisted and rehydrated with this storage.
    */
-  whitelist: ["session", "onboarding", "notifications", "profile", "entities"]
+  whitelist: ["onboarding", "notifications", "profile", "entities"]
 };
 
-const persistedReducer: Reducer<GlobalState, Action> = persistCombineReducers<
-  GlobalState
+const persistedReducer: Reducer<PersistedGlobalState, Action> = persistReducer<
+  GlobalState,
+  Action
 >(persistConfig, rootReducer);
 
 const logger = createLogger({
@@ -66,7 +66,6 @@ function configureStoreAndPersistor(): { store: Store; persistor: Persistor } {
     (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
   const enhancer: StoreEnhancer = composeEnhancers(
     applyMiddleware(
-      thunk,
       sagaMiddleware,
       logger,
       navigation,
@@ -75,7 +74,7 @@ function configureStoreAndPersistor(): { store: Store; persistor: Persistor } {
     )
   );
 
-  const store: Store = createStore<GlobalState, Action, {}, {}>(
+  const store: Store = createStore<PersistedGlobalState, Action, {}, {}>(
     persistedReducer,
     enhancer
   );
