@@ -80,7 +80,7 @@ export const WalletStyles = StyleSheet.create({
     backgroundColor: variables.brandDarkGray
   },
   whiteContent: {
-    backgroundColor: variables.brandGray,
+    backgroundColor: variables.colorWhite,
     flex: 1
   },
   pfImage: {
@@ -92,6 +92,7 @@ export const WalletStyles = StyleSheet.create({
     // temporary -- will be updated with actual component in future version
     height: 120,
     marginLeft: -200,
+    bottom: -10,
     resizeMode: "contain"
   },
   pfTabCard: {
@@ -138,5 +139,10 @@ export const WalletStyles = StyleSheet.create({
     marginTop: 6,
     fontSize: variables.fontSize1,
     color: variables.brandPrimary
+  },
+  walletBannerText: {
+    height: 50,
+    alignItems: "flex-end",
+    flexDirection: "row"
   }
 });
