@@ -16,106 +16,161 @@ import {
 import * as React from "react";
 import { NavigationScreenProp, NavigationState } from "react-navigation";
 
+import { Content } from "native-base";
+import { connect, Dispatch } from "react-redux";
 import I18n from "../../i18n";
 import ROUTES from "../../navigation/routes";
+
+import { selectTransactionForDetails } from "../../store/actions/wallet/transactions";
+import { GlobalState } from "../../store/reducers/types";
+import {
+  latestTransactionsSelector,
+  transactionsByCardSelector
+} from "../../store/reducers/wallet/transactions";
 import { WalletTransaction } from "../../types/wallet";
 import { WalletStyles } from "../styles/wallet";
 
-type Props = Readonly<{
-  title: string;
-  totalAmount: string;
-  navigation: NavigationScreenProp<NavigationState>;
+type ReduxMappedStateProps = Readonly<{
   transactions: ReadonlyArray<WalletTransaction>;
 }>;
 
-type State = Readonly<{
-  data: ReadonlyArray<WalletTransaction>;
+type ReduxMappedDispatchProps = Readonly<{
+  selectTransaction: (i: WalletTransaction) => void;
 }>;
 
 /**
+ * The type of transactions that are to be shown
+ */
+export enum TransactionsDisplayed {
+  LATEST, // show the latest transactions
+  BY_CARD // show all the transactions paid with an already-selected credit card (available in the store)
+}
+
+type OwnProps = Readonly<{
+  title: string;
+  totalAmount: string;
+  navigation: NavigationScreenProp<NavigationState>;
+  display: TransactionsDisplayed;
+}>;
+
+type Props = OwnProps & ReduxMappedStateProps & ReduxMappedDispatchProps;
+/**
  * Transactions List component
  */
-export class TransactionsList extends React.Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = { data: props.transactions };
-  }
-
+class TransactionsList extends React.Component<Props> {
   private renderDate(transaction: WalletTransaction) {
     const datetime: string = `${transaction.date} - ${transaction.time}`;
-    if (transaction.isNew) {
-      return (
-        <Row>
-          <Icon
-            type="FontAwesome"
-            name="circle"
-            active={true}
-            style={WalletStyles.newIconStyle}
-          />
-          <Text note={true}>{datetime}</Text>
-        </Row>
-      );
-    }
     return (
       <Row>
-        <Text note={true}>{datetime}</Text>
+        <Left>
+          <Text>
+            {transaction.isNew && (
+              <Icon
+                type="FontAwesome"
+                name="circle"
+                active={true}
+                style={WalletStyles.newIconStyle}
+              />
+            )}
+            <Text note={true}>
+              {transaction.isNew ? `  ${datetime}` : datetime}
+            </Text>
+          </Text>
+        </Left>
       </Row>
     );
   }
 
-  public render(): React.ReactNode {
-    const { navigate } = this.props.navigation;
-    const ops = this.state.data;
+  private renderRow = (item: WalletTransaction): React.ReactElement<any> => (
+    <ListItem
+      onPress={() => {
+        this.props.selectTransaction(item);
+        this.props.navigation.navigate(ROUTES.WALLET_TRANSACTION_DETAILS);
+      }}
+    >
+      <Body>
+        <Grid>
+          {this.renderDate(item)}
+          <Row>
+            <Left>
+              <Text>{item.paymentReason}</Text>
+            </Left>
+            <Right>
+              <Text>
+                {item.amount} {item.currency}
+              </Text>
+            </Right>
+          </Row>
+          <Row>
+            <Left>
+              <Text note={true}>{item.recipient}</Text>
+            </Left>
+          </Row>
+        </Grid>
+      </Body>
+    </ListItem>
+  );
 
-    if (ops.length === 0) {
-      return <Text>{I18n.t("wallet.noTransactions")}</Text>;
+  public render(): React.ReactNode {
+    const { transactions } = this.props;
+
+    if (transactions.length === 0) {
+      return (
+        <Content scrollEnabled={false} style={WalletStyles.whiteContent}>
+          <Text>{I18n.t("wallet.noTransactions")}</Text>
+        </Content>
+      );
     }
     // TODO: onPress should redirect to the transaction details @https://www.pivotaltracker.com/story/show/154442946
     return (
-      <Grid>
-        <Row>
-          <Left>
-            <Text bold={true}>{this.props.title}</Text>
-          </Left>
-          <Right>
-            <Text>{this.props.totalAmount}</Text>
-          </Right>
-        </Row>
-        <Row>
-          <List
-            removeClippedSubviews={false}
-            dataArray={ops as any[]} // tslint:disable-line
-            renderRow={(item): React.ReactElement<any> => (
-              <ListItem
-                onPress={(): boolean =>
-                  navigate(ROUTES.WALLET_TRANSACTION_DETAILS, {
-                    transaction: item
-                  })
-                }
-              >
-                <Body>
-                  <Grid>
-                    {this.renderDate(item)}
-                    <Row>
-                      <Left>
-                        <Text>{item.paymentReason}</Text>
-                      </Left>
-                      <Right>
-                        <Text>
-                          {item.amount} {item.currency}
-                        </Text>
-                      </Right>
-                    </Row>
-                    <Row>
-                      <Text note={true}>{item.location}</Text>
-                    </Row>
-                  </Grid>
-                </Body>
-              </ListItem>
-            )}
-          />
-        </Row>
-      </Grid>
+      <Content scrollEnabled={false} style={WalletStyles.whiteContent}>
+        <Grid>
+          <Row>
+            <Left>
+              <Text bold={true}>{this.props.title}</Text>
+            </Left>
+            <Right>
+              <Text>{this.props.totalAmount}</Text>
+            </Right>
+          </Row>
+          <Row>
+            <List
+              scrollEnabled={false}
+              removeClippedSubviews={false}
+              dataArray={transactions as WalletTransaction[]} // tslint:disable-line: readonly-array
+              renderRow={this.renderRow}
+            />
+          </Row>
+        </Grid>
+      </Content>
     );
   }
 }
+
+const mapStateToProps = (
+  state: GlobalState,
+  props: OwnProps
+): ReduxMappedStateProps => {
+  switch (props.display) {
+    case TransactionsDisplayed.LATEST: {
+      return {
+        transactions: latestTransactionsSelector(state)
+      };
+    }
+    case TransactionsDisplayed.BY_CARD: {
+      return {
+        transactions: transactionsByCardSelector(state)
+      };
+    }
+  }
+  return { transactions: [] };
+};
+
+const mapDispatchToProps = (dispatch: Dispatch): ReduxMappedDispatchProps => ({
+  selectTransaction: item => dispatch(selectTransactionForDetails(item))
+});
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(TransactionsList);
