@@ -1,11 +1,11 @@
 import { StackNavigator } from "react-navigation";
 import { AddPaymentMethodScreen } from "../screens/wallet/AddPaymentMethodScreen";
-import { CreditCardsScreen } from "../screens/wallet/CreditCardsScreen";
-import { ManuallyIdentifyTransactionScreen } from "../screens/wallet/ManuallyIdentifyTransactionScreen";
-import { TransactionDetailsScreen } from "../screens/wallet/TransactionDetailsScreen";
-import { TransactionsScreen } from "../screens/wallet/TransactionsScreen";
-import { WalletHomeScreen } from "../screens/wallet/WalletHomeScreen";
+import {CreditCardsScreen} from "../screens/wallet/CreditCardsScreen";
+import TransactionDetailsScreen from "../screens/wallet/TransactionDetailsScreen";
+import TransactionsScreen from "../screens/wallet/TransactionsScreen";
+import WalletHomeScreen from "../screens/wallet/WalletHomeScreen";
 import ROUTES from "./routes";
+import { ManuallyIdentifyTransactionScreen } from "../screens/wallet/ManuallyIdentifyTransactionScreen";
 
 /**
  * TODO: migrate WALLET_TRANSACTION_SUMMARY on a new navigator for screens which does not visualize the footer menu.
