@@ -16,41 +16,28 @@ import { isValid } from "redux-form";
 import SpidInformationForm, {
   FORM_NAME as SPID_INFORMATION_FORM_NAME
 } from "../../components/forms/SpidInformationForm";
+import {
+  ContextualHelpInjectedProps,
+  withContextualHelp
+} from "../../components/helpers/withContextualHelp";
 import AppHeader from "../../components/ui/AppHeader";
-import Modal from "../../components/ui/Modal";
 import I18n from "../../i18n";
-import { GlobalState } from "../../reducers/types";
+import { GlobalState } from "../../store/reducers/types";
 type ReduxMappedProps = {
   isFormValid: boolean;
 };
 type OwnProps = {
   navigation: NavigationScreenProp<NavigationState>;
 };
-type Props = ReduxMappedProps & OwnProps;
-type State = {
-  isTosModalVisible: boolean;
-};
 
-// TODO: replace the contextual help with the approariate
-// component @https://www.pivotaltracker.com/story/show/157874540
+type Props = ReduxMappedProps & OwnProps & ContextualHelpInjectedProps;
+
 /**
  * A screen where the user can insert an email to receive information about SPID.
  */
-class SpidInformationRequestScreen extends React.Component<Props, State> {
-  public state: State = {
-    isTosModalVisible: false
-  };
-
+class SpidInformationRequestScreen extends React.Component<Props> {
   private goBack() {
     this.props.navigation.goBack();
-  }
-
-  private showModal() {
-    this.setState({ isTosModalVisible: true });
-  }
-
-  private hideModal() {
-    this.setState({ isTosModalVisible: false });
   }
 
   public render() {
@@ -88,7 +75,7 @@ class SpidInformationRequestScreen extends React.Component<Props, State> {
           <Text>
             {I18n.t("authentication.spid_information_request.paragraph3")}
           </Text>
-          <Text link={true} onPress={_ => this.showModal()}>
+          <Text link={true} onPress={this.props.showHelp}>
             {I18n.t("authentication.spid_information_request.tosLinkText")}
           </Text>
         </Content>
@@ -103,16 +90,6 @@ class SpidInformationRequestScreen extends React.Component<Props, State> {
             </Text>
           </Button>
         </View>
-        <Modal isVisible={this.state.isTosModalVisible} fullscreen={true}>
-          <View header={true}>
-            <Icon name="cross" onPress={_ => this.hideModal()} />
-          </View>
-          <Content>
-            <H1>{I18n.t("personal_data_processing.title")}</H1>
-            <View spacer={true} large={true} />
-            <Text>{I18n.t("personal_data_processing.content")}</Text>
-          </Content>
-        </Modal>
       </Container>
     );
   }
@@ -124,4 +101,11 @@ const mapStateToProps = (state: GlobalState): ReduxMappedProps => ({
    */
   isFormValid: isValid(SPID_INFORMATION_FORM_NAME)(state)
 });
-export default connect(mapStateToProps)(SpidInformationRequestScreen);
+
+export default connect(mapStateToProps)(
+  withContextualHelp(
+    SpidInformationRequestScreen,
+    I18n.t("personal_data_processing.title"),
+    I18n.t("personal_data_processing.content")
+  )
+);
