@@ -10,16 +10,18 @@ import {
   StoreEnhancer as ReduxStoreEnhancer
 } from "redux";
 
-import { GlobalState } from "../../reducers/types";
+import { GlobalState } from "../reducers/types";
 import { ApplicationActions } from "./application";
 import { APP_STATE_CHANGE_ACTION } from "./constants";
 import { ErrorActions } from "./error";
 import { MessagesActions } from "./messages";
 import { NotificationsActions } from "./notifications";
 import { OnboardingActions } from "./onboarding";
+import { PinloginActions } from "./pinlogin";
 import { ProfileActions } from "./profile";
 import { ServicesActions } from "./services";
 import { SessionActions } from "./session";
+import { WalletActions } from "./wallet";
 
 export type ApplicationState = "background" | "inactive" | "active";
 
@@ -35,17 +37,16 @@ export type Action =
   | SessionActions
   | OnboardingActions
   | NotificationsActions
+  | PinloginActions
   | ProfileActions
   | MessagesActions
   | ServicesActions
-  | ErrorActions;
+  | ErrorActions
+  | WalletActions;
 
-/* eslint-disable no-use-before-define */
-// We need to disable the eslint rule because of a problem described here
-// @https://github.com/babel/babel-eslint/issues/485
 export type GetState = () => GlobalState;
 
-export type Dispatch = DispatchAPI<Action, GlobalState>;
+export type Dispatch = DispatchAPI<Action>;
 
 export type Store = ReduxStore<GlobalState>;
 
