@@ -6,9 +6,10 @@
  */
 
 import { isSome, none, Option } from "fp-ts/lib/Option";
-import { GlobalState } from "../../reducers/types";
+
 import { ERROR_CLEAR, FetchRequestActionsType } from "../actions/constants";
 import { Action } from "../actions/types";
+import { GlobalState } from "./types";
 
 export type ErrorState = Readonly<
   { [key in FetchRequestActionsType]: Option<string> }
