@@ -9,6 +9,12 @@ declare module "native-base" {
     interface Text extends TextProperties {
       link?: boolean;
       bold?: boolean;
+      formatDate?: boolean;
+      leftAlign?: boolean;
+      rightAlign?: boolean;
+      alternativeBold?: boolean;
+      white?: boolean;
+      alignCenter?: boolean;
     }
   }
 }
@@ -22,7 +28,24 @@ export default (): Theme => {
     ".bold": {
       ...makeFontStyleObject(Platform.select, variables.textBoldWeight)
     },
-
-    lineHeight: variables.lineHeight
+    ".leftAlign": {
+      textAlign: "left"
+    },
+    ".rightAlign": {
+      textAlign: "right"
+    },
+    ".alternativeBold": {
+      lineHeight: variables.lineHeight1,
+      fontWeight: variables.textBoldWeight,
+      color: variables.h1Color,
+      ...makeFontStyleObject(Platform.select, variables.textBoldWeight)
+    },
+    ".white": {
+      color: variables.colorWhite
+    },
+    ".alignCenter": {
+      textAlign: "center"
+    },
+    lineHeight: variables.lineHeightBase
   };
 };
